created Statistics - created Statistics class, which saves basic table statistics including #rows, column sizes, column maximums and minimums, and column null data - updated subclasses of Data class - updated creation of each data class when reading records - ScanNode now recognizes basic table statistics while reading CSV

+++ wolverine-master/src/main/java/table/Data/FloatData.java
@@ -1,6 +1,7 @@
 package table.Data;
 
 public class FloatData extends Data {
+
     private float value;
 
     public FloatData(String value) {
@@ -11,11 +12,12 @@ public class FloatData extends Data {
         return value;
     }
 
-    public void setValue(int value) {
+    public void setValue(float value) {
         this.value = value;
     }
 
     public String toString() {
         return String.format("%f", value);
     }
+
 }
